rename paramBox templates
