+++ arknights-tracker/src/lib/components/dataToolbarV2/sortDropdowns/RecipesSortDropdown.svelte
@@ -3,7 +3,7 @@
     import DropdownTemplate from "$lib/components/dataToolbarV2/DropdownTemplate.svelte";
     import AlphabeticSortSelector from "$lib/components/dataToolbarV2/sortDropdowns/AlphabeticSortSelector.svelte";
     import RarityParamBox from "$lib/components/dataToolbarV2/paramBoxes/RarityParamBox.svelte";
-    import TextParamBox from "$lib/components/dataToolbarV2/paramBoxes/TextParamBox.svelte";
+    import TextParamBoxTemplate from "$lib/components/dataToolbarV2/paramBoxes/TextParamBoxTemplate.svelte";
     import DraggableParamList from "$lib/components/dataToolbarV2/sortDropdowns/DraggableParamList.svelte";
     import DraggableSortGroups from "$lib/components/dataToolbarV2/sortDropdowns/DraggableSortGroups.svelte";
     import { t } from "$lib/i18n";
@@ -87,7 +87,7 @@
         {:else}
 
             <DraggableParamList
-                paramBox={TextParamBox}
+                paramBox={TextParamBoxTemplate}
                 getLocaleFunc={(param) => getFilterNameLocale(openedSortField, param)}
                 bind:paramList={sortParams.sortFieldParams[openedSortField]}
             />
